apply patch for warning removal in tiles code
Author: Kiran Misra <misrak@sharplabs.com>


git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/branches/HM-4.1-dev@1464 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/Lib/TLibDecoder/NALread.cpp
@@ -63,7 +63,7 @@ static void convertPayloadToRBSP(vector<uint8_t>& nalUnitBuf)
       if ( (*(it_read) == 0x00) && (*(it_read+1) == 0x00) && (*(it_read+2) == 0x02) ) // tile marker detected
       {
         it_read += 2;
-        UInt uiDistance  = it_write - nalUnitBuf.begin();
+        UInt uiDistance  = (UInt) (it_write - nalUnitBuf.begin());
         UInt uiCount     = pcBitstream->getTileMarkerLocationCount();
         bTileMarkerFound = true;
         pcBitstream->setTileMarkerLocation( uiCount, uiDistance );
@@ -88,7 +88,7 @@ static void convertPayloadToRBSP(vector<uint8_t>& nalUnitBuf)
     {
 #if TILES_DECODER
       // update tile marker location
-      UInt uiDistance = it_read - nalUnitBuf.begin();
+      UInt uiDistance = (UInt) (it_read - nalUnitBuf.begin());
       for (UInt uiIdx=0; uiIdx<pcBitstream->getTileMarkerLocationCount(); uiIdx++)
       {
         if (auiStoredTileMarkerLocation[ uiIdx ] >= uiDistance)
